finished account management

+++ src/features/movies/components/LogoutModal.jsx
@@ -1,25 +1,91 @@
 import "./LogoutModal.scss";
 import { closeLogin } from "../modalSlice";
-import { setLogIn, clearToken } from "../accountSlice";
+import {
+  setLogIn,
+  clearToken,
+  checkDelete,
+  deleteConfirm,
+  selectPassword,
+  setPassword,
+  setMessage,
+  selectMessage,
+  clearInputs,
+  changePassword,
+  checkPasswordChange,
+  setNewPassword,
+  selectPassword2,
+  selectPassword3,
+  confirmNewPassword,
+} from "../accountSlice";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteUser } from "../../../utils/apis";
+import { deleteUser, updatePassword } from "../../../utils/apis";
 import axios from "axios";
 import { selectToken } from "../accountSlice";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 const LogoutModal = () => {
   const dispatch = useDispatch();
   const token = useSelector(selectToken);
+  const isDeleteClicked = useSelector(checkDelete);
+  const password = useSelector(selectPassword);
+  const message = useSelector(selectMessage);
+  const changePasswordSelected = useSelector(checkPasswordChange);
+  const password2 = useSelector(selectPassword2);
+  const password3 = useSelector(selectPassword3);
+
+  const notifyPswChg = () => toast(`Password Successfully Changed!`);
+
+  const passwordChange = async () => {
+    if (password2 !== password3) {
+      dispatch(setMessage("Passwords do not match"));
+      return;
+    }
+    if (!password2 || !password3) {
+      dispatch(setMessage("Enter new Password"));
+      return;
+    }
+    if (password2 === password) {
+      dispatch(setMessage("New Password must be different"));
+      return;
+    }
+    const api = updatePassword(token, password, password2);
+    try {
+      const updateResult = await axios.patch(api);
+      const updateStatus = updateResult.data.status;
+      if (updateStatus === 0) {
+        dispatch(setMessage("Password Incorrect"));
+      }
+
+      if (updateStatus === 1) {
+        notifyPswChg();
+        dispatch(changePassword());
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   const deleteAccount = async () => {
-    const api = deleteUser(token);
+    if (!password || password.length === 0) {
+      const message = "Enter Password";
+      dispatch(setMessage(message));
+      return;
+    }
+    const api = deleteUser(token, password);
     try {
       const deleteResult = await axios.delete(api);
       const deleteStatus = deleteResult.data.status;
       console.log(deleteResult.data.status);
-      //   dispatch(setMessage(loginStatus));
+      if (deleteStatus === 0) {
+        dispatch(setMessage("Password Incorrect"));
+      }
+
       if (deleteStatus === 1) {
         dispatch(setLogIn());
         dispatch(closeLogin());
+        dispatch(clearInputs());
+        dispatch(deleteConfirm());
       }
     } catch (error) {
       console.log(error);
@@ -28,6 +94,7 @@ const LogoutModal = () => {
 
   return (
     <>
+      <ToastContainer />
       <aside className="logoutModalContainer">
         <div className="modal">
           <div className="btnContainer">
@@ -36,6 +103,13 @@ const LogoutModal = () => {
               className="exitButton"
               onClick={() => {
                 dispatch(closeLogin());
+                dispatch(clearInputs());
+                if (isDeleteClicked) {
+                  dispatch(deleteConfirm());
+                }
+                if (changePasswordSelected) {
+                  dispatch(changePassword());
+                }
               }}
             >
               X
@@ -53,6 +127,13 @@ const LogoutModal = () => {
                   dispatch(setLogIn());
                   dispatch(closeLogin());
                   dispatch(clearToken());
+                  dispatch(clearInputs());
+                  if (isDeleteClicked) {
+                    dispatch(deleteConfirm());
+                  }
+                  if (changePasswordSelected) {
+                    dispatch(changePassword());
+                  }
                 }}
               >
                 Log Out
@@ -62,51 +143,116 @@ const LogoutModal = () => {
               <button
                 type="submit"
                 className="changePasswordButton"
-                //   onClick={() => {
-                //     dispatch(setLogIn());
-                //     dispatch(closeLogin());
-                //     dispatch(clearToken());
-                //   }}
+                onClick={() => {
+                  dispatch(changePassword());
+                  dispatch(clearInputs());
+                }}
+                style={
+                  isDeleteClicked
+                    ? { pointerEvents: "none" }
+                    : { pointerEvents: "auto" }
+                }
               >
-                Change Password
+                {changePasswordSelected ? "Cancel" : "Change Password"}
               </button>
             </div>
+            {changePasswordSelected ? (
+              <div className="passwordChangeDiv">
+                <label htmlFor="psw1" className="psw1Label">
+                  Current Password
+                </label>
+                <input
+                  value={password || ""}
+                  type="password"
+                  placeholder="Enter Current Password"
+                  name="psw"
+                  required
+                  className="enterOldPswInput"
+                  onInput={(e) => dispatch(setPassword(e.target.value))}
+                />
+                <label htmlFor="psw2" className="psw2Label">
+                  New Password
+                </label>
+                <input
+                  value={password2 || ""}
+                  type="password"
+                  placeholder="Enter New Password"
+                  name="psw2"
+                  required
+                  className="enterNewPswInput"
+                  onInput={(e) => dispatch(setNewPassword(e.target.value))}
+                />
+                <label htmlFor="psw3" className="psw3Label">
+                  Confirm New Password
+                </label>
+                <input
+                  value={password3 || ""}
+                  type="password"
+                  placeholder="Confirm New Password"
+                  name="psw3"
+                  required
+                  className="confirmNewPswInput"
+                  onInput={(e) => dispatch(confirmNewPassword(e.target.value))}
+                />
+                <p className="message">{message}</p>
+                <div className="confirmChangePswButtonDiv">
+                  <button
+                    type="submit"
+                    className="confirmChangePswButton"
+                    onClick={() => {
+                      passwordChange();
+                    }}
+                  >
+                    Change Password
+                  </button>
+                </div>
+              </div>
+            ) : null}
             <div className="deleteAccountButtonDiv">
               <button
                 type="submit"
                 className="deleteAccountButton"
-                // onClick={() => {
-                //   deleteAccount();
-                //   // dispatch(setLogIn());
-                //   // dispatch(closeLogin());
-                //   // dispatch(clearToken());
-                // }}
-                onClick={deleteAccount}
+                onClick={() => {
+                  dispatch(deleteConfirm());
+                  dispatch(clearInputs());
+                }}
+                style={
+                  changePasswordSelected
+                    ? { pointerEvents: "none" }
+                    : { pointerEvents: "auto" }
+                }
               >
-                Delete Account
+                {isDeleteClicked ? "Cancel" : "Delete Account"}
               </button>
             </div>
-            <div className="checkDeleteTitle">
-              <p>Enter password to confirm account deletion</p>
-              <input
-                // value={password || ""}
-                type="password"
-                placeholder="Enter Password"
-                name="psw"
-                required
-                className="pswDeleteInput"
-                // onInput={(e) => dispatch(setPassword(e.target.value))}
-              />
-              <div className="deleteButtonDiv">
-                <button
-                  type="submit"
-                  className="deleteAccountConfirmButton"
-                  //   onClick={register}
-                >
-                  Delete Account
-                </button>
+            {isDeleteClicked ? (
+              <div className="deleteConfirmDiv">
+                <div className="checkDeleteTitle">
+                  <p>Enter password to confirm account deletion</p>
+                </div>
+                <div className="pswDeleteInputDiv">
+                  <input
+                    value={password || ""}
+                    type="password"
+                    placeholder="Enter Password"
+                    name="psw"
+                    required
+                    className="pswDeleteInput"
+                    onInput={(e) => dispatch(setPassword(e.target.value))}
+                  />
+                </div>
+                <p className="message">{message}</p>
+                <div className="deleteButtonDiv">
+                  <button
+                    type="submit"
+                    className="deleteAccountConfirmButton"
+                    onClick={deleteAccount}
+                  >
+                    Delete Account
+                  </button>
+                </div>
               </div>
-            </div>
+            ) : null}
           </div>
         </div>
       </aside>
